add a support that it can be used in Java source code.
We can specify the analysis target in the Java program
and get the cyclomatic number by invoking "getCyclomaticNumber" method.
Previously, this tool was supposed to be used on the command line.
This commit has expanded the possibilities.

Specifically, we make two major changes.
First, we made the "count" variable a non-static field.
In the second, we added a getter method that returns a cyclomatic
number.